+++ tests/plugin_test/src/164_array_values_redundant.php
@@ -1,0 +1,7 @@
+<?php
+function test_array_values(...$args) {
+    $arr = array_values($args);
+    sort($arr);
+    return array_values($arr);
+}
+test_array_values(1,2);